Update BlueTooth_orderer_3Effects_SimplerBACKUP.ino
added code for another pushbutton and tried implementing a bypass setting where all effects are disconnected and just a dry signal is run to the output. Also changed how the effect parameters are edited. Instead of changing on each turn, the variable now updates and the effect takes placed once you push the pushbutton to exit like you would previously.